--- Web/starter_project_vanguard/components/blog/Image.tsx
@@ -1,22 +1,0 @@
-import React, { ReactNode } from 'react';
-
-interface ImageProps {
-  src: string;
-  borderRadius?: string;
-  className?: string;
-}
-
-const Image: React.FC<ImageProps> = ({
-  src,
-  borderRadius = 'rounded',
-  className,
-}) => {
-  return (
-    <div className={className}>
-      <img src={src} className={`object-cover ${borderRadius}`} />
-      
-    </div>
-  );
-};
-
-export default Image;
