--- Web/starter_project_vanguard/components/blog/ProfilePic.tsx
@@ -1,16 +1,0 @@
-import React from 'react';
-import Avatar from './Avatar';
-
-function ProfilePic() {
-  return (
-    <div>
-     
-      <Avatar 
-      src="https://media.wired.com/photos/598e35fb99d76447c4eb1f28/16:9/w_2123,h_1194,c_limit/phonepicutres-TA.jpg" 
-      alt="Avatar" 
-      size="md" />
-    </div>
-  );
-}
-
-export default ProfilePic;
