+++ Web/starter_project_vanguard/components/common/Avatar.tsx
