+++ Web/starter_project_vanguard/components/common/Text.tsx
